(pd) move collision fns into utils
